use dynamicsimulation in package naming
