+++ dynamic-simulation/dynamic-simulation-api/src/main/java/com/powsybl/dynamicsimulation/json/DynamicSimulationParametersDeserializer.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynamic.simulation.json;
+package com.powsybl.dynamicsimulation.json;
 
 import java.io.IOException;
 import java.util.Collections;
@@ -16,7 +16,7 @@ import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 import com.powsybl.commons.extensions.Extension;
 import com.powsybl.commons.json.JsonUtil;
-import com.powsybl.dynamic.simulation.DynamicSimulationParameters;
+import com.powsybl.dynamicsimulation.DynamicSimulationParameters;
 
 /**
  * @author Marcos de Miguel <demiguelm at aia.es>
