+++ dynamic-simulation/dynamic-simulation-api/src/main/java/com/powsybl/dynamicsimulation/json/JsonDynamicSimulationParameters.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.dynamic.simulation.json;
+package com.powsybl.dynamicsimulation.json;
 
 import java.io.IOException;
 import java.io.InputStream;
@@ -26,7 +26,7 @@ import com.powsybl.commons.extensions.Extension;
 import com.powsybl.commons.extensions.ExtensionJsonSerializer;
 import com.powsybl.commons.extensions.ExtensionProviders;
 import com.powsybl.commons.json.JsonUtil;
-import com.powsybl.dynamic.simulation.DynamicSimulationParameters;
+import com.powsybl.dynamicsimulation.DynamicSimulationParameters;
 
 /**
  * Provides methods to read and write DynamicSimulationParameters from and to JSON.
